Completed

+++ app/Http/Controllers/CategoriesController.php
@@ -98,6 +98,14 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+
+        if($category->count() > 0){
+            
+            session()->flash('error','Category cannot be deleted because it has some posts');
+
+            return redirect()->back();
+        }
+
         $category->delete();
         session()->flash('success','Category Deleted Successfully');
